# docs/openapi/order-service.yaml
openapi: 3.1.0
info:
  title: order-service
  version: 0.0.1
  description: Owns the order lifecycle. Publishes order-created, consumes order-accepted and driver-assigned.

servers:
  - url: http://localhost:8082

paths:
  /health:
    get:
      summary: Liveness probe
      operationId: health
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /metrics-lite:
    get:
      summary: Lightweight operational snapshot for ops-dashboard polling
      operationId: metricsLite
      responses:
        "200":
          description: Current service metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricsLiteResponse"

  /orders:
    post:
      summary: Place a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "202":
          description: Order accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid request

  /orders/{orderId}:
    get:
      summary: Get order status
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Not found

components:
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [UP, DOWN]

    MetricsLiteResponse:
      type: object
      required: [service, status, uptimeSeconds, ordersCreatedTotal, ordersDeliveredTotal]
      properties:
        service:
          type: string
          example: order-service
        status:
          type: string
          enum: [UP, DOWN]
        uptimeSeconds:
          type: integer
        ordersCreatedTotal:
          type: integer
        ordersDeliveredTotal:
          type: integer

    CreateOrderRequest:
      type: object
      required: [customerId, restaurantId, items, deliveryAddress]
      properties:
        customerId:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        deliveryAddress:
          $ref: "#/components/schemas/DeliveryAddress"

    OrderItem:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1

    DeliveryAddress:
      type: object
      required: [lat, lng, street, city]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        street:
          type: string
        city:
          type: string

    OrderResponse:
      type: object
      required: [orderId, status, totalCents, createdAt]
      properties:
        orderId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, DRIVER_ASSIGNED, DELIVERED]
        totalCents:
          type: integer
        etaMinutes:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time